Update Config For TaskForm
- Update Suspense for TaskPage to build successfully
